- Added `has_pattern`
